Created library.c [INITIAL COMMIT]
okay so this is the first commit, Everything seems to work. All the functions work. Except issueBook(); I didn't code its logic in. I will do that tomorrow. Other than that, Yeah I mean it works... The code really needs to be cleaned up... If the indentation looks kinda weird... It's codeblocks.